% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_compr}
\alias{data_compr}
\title{Comprehensive example curved data with all segment features present}
\format{
\subsection{\code{data_compr}}{

A data frame with 237 rows and 3 columns:
\describe{
\item{x}{predictor value}
\item{y_true}{the true response value}
\item{y}{the response with noise}
}
}
}
\source{
see data-raw
}
\usage{
data_compr
}
\description{
Practice and example dataset to text package functionality
}
\keyword{datasets}
