% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_multivar}
\alias{data_multivar}
\title{Multivariate example data}
\format{
\subsection{\code{data_multivar}}{

A data frame with 100 rows and 5 columns:
\describe{
\item{x1}{numerical predictor }
\item{x2}{factor predictor }
\item{x3}{numerical predictor }
\item{y_true}{the true response value}
\item{y}{the response with noise}
}
}
}
\source{
see data-raw
}
\usage{
data_multivar
}
\description{
Practice and example dataset to text package functionality
}
\keyword{datasets}
