% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_norm}
\alias{data_norm}
\title{Normally-distributed data}
\format{
\subsection{\code{data_norm}}{

A data frame with 200 rows and 3 columns:
\describe{
\item{x}{predictor value}
\item{y_true}{the true response value}
\item{y}{the response with noise}
}
}
}
\source{
see data-raw
}
\usage{
data_norm
}
\description{
Practice and example dataset following a normal distribution
}
\keyword{datasets}
