% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_sig_sin}
\alias{data_sig_sin}
\title{Sigmoid-sine curve data}
\format{
\subsection{\code{data_sig_sin}}{

A data frame with 101 rows and 3 columns:
\describe{
\item{x}{predictor value}
\item{y_true}{the true response value}
\item{y}{the response with noise}
}
}
}
\source{
see data-raw
}
\usage{
data_sig_sin
}
\description{
Practice and example dataset to text package functionality
}
\keyword{datasets}
