% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_lextrema.R
\name{.check_gam}
\alias{.check_gam}
\title{Check that gam object}
\usage{
.check_gam(mod, var, smooth, multivariate)
}
\arguments{
\item{mod}{gam model object to be evaluated}

\item{var}{predictor variable name}

\item{smooth}{for multivariate models, smooth is the smooth term to be evaluated in the format "s(...)"}

\item{multivariate}{TRUE/FALSE if the model to be evaluated is multivariate or not}
}
\value{
returns the est_slopes object with an added column defining the sections
}
\description{
Evaluates if the gam object is appropriate to be passed on to lextrema
}
