% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_lextrema.R
\name{.eval_segments}
\alias{.eval_segments}
\title{Evaluate  sections}
\usage{
.eval_segments(est_slopes)
}
\arguments{
\item{est_slopes}{gam model object to be evaluated}
}
\value{
returns the est_slopes object with an added column defining the sections
}
\description{
Evaluates sections identified by find_lextrema (confidence interval of first derivative does not exclude 0)  to identify it as a peak, trough, upper step or lower step
}
