% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_lextrema.R
\name{.plot_lextrema_multivar}
\alias{.plot_lextrema_multivar}
\title{Plot the lextrema multivariate}
\usage{
.plot_lextrema_multivar(
  lextr,
  plot_deriv = TRUE,
  show_segs = c("local_max", "local_min"),
  show_segs_deriv = TRUE,
  type = "response"
)
}
\arguments{
\item{lextr}{a lextrema object produced by the function quantify_lextrema}

\item{plot_deriv}{boolean TRUE/FALSE if the first derivative should be plotted}

\item{show_segs}{which segment types to show. see lextremagam::feature_codes for the options}

\item{show_segs_deriv}{TRUE/FALSE if the derivative plot should be produced}

\item{type}{if the data should be plotted on the repsonse scale (default) or link scale}
}
\value{
a ggplot object
}
\description{
Returns a plot object
}
