% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_lextrema.R
\name{.quantify_lextrema_mutlivar}
\alias{.quantify_lextrema_mutlivar}
\title{Estimate the first derivative}
\usage{
.quantify_lextrema_mutlivar(
  mod,
  var = NULL,
  smooth = NULL,
  step_size = NULL,
  conf_level = 0.7763932,
  deriv_method = c("gratia", "marginaleffects")
)
}
\arguments{
\item{mod}{gam model object to be evaluated}

\item{var}{predictor variable name to be evaluated}

\item{smooth}{smooth name to be evaluated in the format \"s()\" (ex: \"s(temp)\")}

\item{step_size}{the step size at which to evaluate the first derivative}

\item{conf_level}{the confidence level (between 0 and 1) at which the confidence interval of the first derivative is estimated}

\item{deriv_method}{whether to use gratia's derivatives function or marginaleffects' slopes function}
}
\value{
a list object built by marginaleffects::slopes or gratia::derivative, including the model rowid, term, estimate, std.error, conf.low, conf.high, y, x
}
\description{
Evaluates the first derivative at a given (very small) step size for a given gam model input. Note: THIS FUNCTION IS CURRENTLY UNIVARIATE
}
