% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_codes}
\alias{feature_codes}
\title{List of approved codes for segment features}
\format{
\subsection{\code{feature_codes}}{

A vector with 8 possible values
}
}
\source{
see data-raw
}
\usage{
feature_codes
}
\description{
Reference list for all possible segment feature codes
}
\keyword{datasets}
