% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_lextrema.R
\name{find_segments}
\alias{find_segments}
\title{Identify the local extremas}
\usage{
find_segments(est_slopes, var, deriv_method)
}
\arguments{
\item{est_slopes}{an output of the marginaleffects::slopes function, including the row ID, the x and y values, the conf.low and conf.high}

\item{var}{predictor variable name}

\item{deriv_method}{gratia or marginaleffects}
}
\description{
Identify the local extremas
}
