% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lextremagam-package.R
\docType{package}
\name{lextremagam}
\alias{lextremagam-package}
\alias{lextremagam}
\title{lextremagam}
\description{
The lextremagam package serves as a streamlined way to quantify peaks and troughs in complex data using generalized additive models. While it was developed primarily for ecologists, it can be applied to any dataset where a local extremum (peak or trough) in a one-dimentional smooth is of interest. Broadly, this package takes a gam object generated using mgcv and identifies confidence intervals for local extrema (peaks or troughs) based on the first derivative of the smooth. This method, its application and its performance are described in [Dupont et al. preprint]. lextremagam can also identify other segments (increasing, decreasing and plateaus) based on the first derivative, but the properties of these segments have not been evaluated and they must be interpreted with caution.
}
\author{
Natalie Dupont

Eric Pedersen
}
\keyword{internal}
