% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_test.R
\name{peaktest_quad}
\alias{peaktest_quad}
\title{Tests for the presence of a peak using the quadratic test. \strong{\strong{NOT AN APPRORPIATE APPROACH}}}
\usage{
peaktest_quad(x, y)
}
\arguments{
\item{x}{vector of x-values from the data}

\item{y}{vector of y-values from the data}
}
\value{
dasf
}
\description{
function from Uri Simonsohn 2018 (https://osf.io/wdbmr)
}
