% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_lextrema.R
\name{plot_lextrema}
\alias{plot_lextrema}
\title{Plot the lextrema}
\usage{
plot_lextrema(
  lextr,
  plot_deriv = TRUE,
  show_segs = c("local_max", "local_min"),
  show_segs_deriv = TRUE,
  type = "response"
)
}
\arguments{
\item{lextr}{a lextrema object produced by the function quantify_lextrema}

\item{plot_deriv}{boolean TRUE/FALSE if the first derivative should be plotted}

\item{show_segs}{which segment types to show. see lextremagam::feature_codes for the options}

\item{show_segs_deriv}{TRUE/FALSE if the derivative plot should show segments}

\item{type}{on what scale to draw the data. The options are "response" or "link" with "response" as the default}
}
\value{
one or two ggplot object
}
\description{
Returns a plot showing delineated segments along a smooth when given a lextrema object. The lextrema object can be univariate or multivariate. Can also provide the first derivative plot of the smooth.
}
