% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_lextrema.R
\name{quantify_lextrema}
\alias{quantify_lextrema}
\title{Estimate the first derivative}
\usage{
quantify_lextrema(model, step_size = NULL, conf_level = 0.95)
}
\arguments{
\item{model}{gam model object to be evaluated}

\item{step_size}{the step size at which to evaluate the first derivative}

\item{conf_level}{the confidence level (between 0 and 1) at which the confidence interval of the first derivative is estimated}
}
\value{
a slopes object dataframe built by marginaleffects::slopes, including the model rowid, term, estimate, std.error, conf.low, conf.high, y, x
}
\description{
Evaluates the first derivative at a given (very small) step size for a given gam model input. Note: THIS FUNCTION IS CURRENTLY UNIVARIATE
}
