% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_lines.R
\name{peaktest_twolines}
\alias{peaktest_twolines}
\title{Detecting the presence of a peak using the two lines test}
\usage{
peaktest_twolines(x, y)
}
\arguments{
\item{x}{vector of x-values (predictor)}

\item{y}{vector of y-values (response)}
}
\value{
a slopes object dataframe built by marginaleffects::slopes, including the model rowid, term, estimate, std.error, conf.low, conf.high, y, x
}
\description{
function from Uri Simonsohn 2018 (https://osf.io/wdbmr)
}
