% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_lextrema.R
\name{plot_lextrema}
\alias{plot_lextrema}
\title{Plot the lextrema (STILL UNIVARIATE)}
\usage{
plot_lextrema(
  quant_segments,
  plot_deriv = TRUE,
  show_segs = c("peak", "trough")
)
}
\arguments{
\item{quant_segments}{a lextrema object produced by the function quantify_lextrema}

\item{plot_deriv}{boolean TRUE/FALSE if the first derivative should be plotted}
}
\value{
a ggplot object
}
\description{
Returns a plot object
}
