% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_lextrema.R
\name{quantify_lextrema}
\alias{quantify_lextrema}
\title{Identify local maxima and minima in a smooth from a GAM}
\usage{
quantify_lextrema(
  mod,
  var = NULL,
  step_size = NULL,
  conf_level = 0.95,
  deriv_method = c("gratia", "marginaleffects"),
  frequentist = FALSE
)
}
\arguments{
\item{mod}{gam model object to be evaluated}

\item{var}{predictor variable over which the slope is evaluated}

\item{step_size}{the step size at which to evaluate the first derivative}

\item{conf_level}{the confidence level (between 0 and 1) at which the confidence interval of the first derivative is estimated}

\item{deriv_method}{whether to use gratia's derivatives function or marginaleffects' slopes function}

\item{frequentist}{if using gratia::derivatives, TRUE/FALSE boolean whether or not the derivatives are calculated using the bayesian (default) or frequentist covariance matrix}
}
\value{
Returns a lextrema object, which includes a model_slopes table, a segment summary table, the input gam model object, and the input var and deriv_method variables
}
\description{
Evaluates the first derivative at a given (very small) step size along the smooth of a given gam model input.
}
\details{
Note: THIS FUNCTION IS ONLY FOR UNIVARIATE GAMs. For multivariate GAMs (ie, multiple predictors), please use quantify_lextrema_multivar()
}
