% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_lines.R
\name{reg2}
\alias{reg2}
\title{Fits the two-line regression based on a given break point xc}
\usage{
reg2(x, y, xc)
}
\arguments{
\item{x}{vector of x-values from the data}

\item{y}{vector of y-values from the data}

\item{xc}{the step size at which to evaluate the first derivative}
}
\value{
a slopes object dataframe built by marginaleffects::slopes, including the model rowid, term, estimate, std.error, conf.low, conf.high, y, x
}
\description{
reg2 function from Uri Simonsohn 2018 (https://osf.io/wdbmr)
}
